myproject/requirements.txt
Django>=4.2
djangorestframework
plotly
pandas